Dynamic css creation moved to server
The jQuery dynamic css caused a flicker in the game as it was being processed every turn. The css is now created once on initial play by the server and served as a file. Ive also added normalize.css to ensure some kind of uniformity across browsers, though firefox is still adding height to the table-cell elements for some reason.